Try to find "good" "skeletal muscle" biosamples
find_releated_terms is some prototype code looking for
related terms in the ontologies the ENCODE DCC is using.

find_flags finds all the biosamples & experiments tagged with the
biosample_term_ids found with find_releated_terms, and then goes
and grabs the audit flags from the encodeproject.org site.

Since the audits are generated on the website and aren't actually in
the objects.