try using server_lifestyle hooks to load the gtf file.
This snapshot is wrong, as main has been flailed on after the hook
didn't actually manage to load the gtf file.

I think this version of main is after I decided to try
and just put the fully created data structure into the .h5 file
instead of reading and parsing the gtf file.

Since it seemed like each request was going to go reread the file